MetaPhylo

Documentation for MetaPhylo.

AbstractTrees.treebreadth

AbstractTrees.treeheight

AbstractTrees.treesize

MetaPhylo.ancestors

MetaPhylo.common_ancestor

MetaPhylo.distance

MetaPhylo.distance_matrix

MetaPhylo.isinternal

MetaPhylo.isinternal

MetaPhylo.isleaf

MetaPhylo.isleaf

MetaPhylo.ladderize!

MetaPhylo.leafedges

MetaPhylo.leaves

MetaPhylo.parent_branch

MetaPhylo.reindex!

MetaPhylo.reroot!

MetaPhylo.swap!

MetaPhylo.swapchildren!

MetaPhylo.treelength

AbstractTrees.treebreadth — Method

AbstractTrees.treebreadth(tree::Tree)

Return the number of leaves in the tree.

AbstractTrees.treeheight — Method

AbstractTrees.treeheight(tree::Tree)

Return the maximum depth from the root to the leaves in the tree. See also treelength.

AbstractTrees.treesize — Method

AbstractTrees.treesize(tree::Tree)

Return the size og the tree.

MetaPhylo.ancestors — Method

ancestors(tree::Tree, idx::Integer)

Return the indices of all ancestor nodes of the specified idx node.

MetaPhylo.common_ancestor — Method

common_ancestor(tree::Tree, idx1::Integer, idx2::Integer)

Return the common ancestor index of two specified idx1 and idx2 nodes.

MetaPhylo.distance — Method

distance(tree::Tree, idx1::Integer, idx2::Integer)

Return distance between two nodes on a tree. The Tree branch types must have the Length trait.

MetaPhylo.distance_matrix — Method

distance_matrix(tree::Tree)

Return pairwise distances between all leaves on the tree in a AxisArray. The Tree branch types must have the Length trait.

MetaPhylo.isinternal — Method

isinternal(tree::Tree, idx::Integer)

Return true if the edge is both connected to internal nodes of the tree.

MetaPhylo.isinternal — Method

isinternal(tree::Tree, idx::Integer)

Return true if the idx is contained in an internal node of the tree.

MetaPhylo.isleaf — Method

isleaf(tree::Tree, edge::Edge)

Return true if the edge is connected to a leaf node of the tree.

MetaPhylo.isleaf — Method

isleaf(tree::Tree, idx::Integer)

Return true if the idx is contained in a leaf node of the tree.

MetaPhylo.ladderize! — Method

ladderize!(tree::Tree; left=false)

Ladderize the tree structure. By default, the smallest clade is on the right side; if left=true, on the left side.

MetaPhylo.leafedges — Method

leafedges(tree::Tree, [idx::Integer])

Return the edges of all edges connected to the leaves in the tree. If the index is specified, this returns the edges connected to its leaves in the tree.

MetaPhylo.leaves — Method

leaves(tree::Tree, [idx::Integer])

Return the indices of all leaves in the tree. If the index is specified, this returns the indices of leaves in its subtree.

MetaPhylo.parent_branch — Method

parent_branch(tree::Tree, idx::Integer)

Return the baranch (edge) between the specified idx node and its parent node. If the node is root, this returns nothing

MetaPhylo.reindex! — Method

reindex!(tree::Tree)

Reindex the tree in PreOderDFS order from the root.

MetaPhylo.reroot! — Method

reroot!(tree::Tree, idx::Integer)

Reroot the tree at the specified node. Return true if rerooting success.

MetaPhylo.swap! — Method

swap!(tree::Tree, idx::Integer, old_new::Pair{<:Integer, <:Integer})

Swap the two child elements of the specified idx node. The old and new in old_new must be child of idx node. Return ture if swapping fails; true otherwise.

MetaPhylo.swapchildren! — Method

swapchildren!(tree::Tree, idx::Integer, newchildren)

Swap the child indices of the specified idx node to the given newchildren. The elements of children and newchildren must be match. Return false if swapping fails; true otherwise.

MetaPhylo.treelength — Method

treelength(tree::Tree, [idx::Integer])

Return maximum distance from the root to the leaves in the tree. If the index is specified, this returns maximum distance from the specified idx to its leaves in the tree. The Tree branch types must have the Length trait. See also treeheight.